requirements.txt
flask==2.3.3
werkzeug==2.3.7
transformers==4.30.2
torch==2.0.1
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
spacy==3.5.3
sentencepiece==0.1.99
protobuf==3.20.3
accelerate==0.20.3 